精简项目+router调整+eslint调整

+++ webpage/4_vue-element-admin/.eslintrc.js
@@ -162,7 +162,7 @@ module.exports = {
       }
     }],
     'padded-blocks': [2, 'never'],
-    'quotes': [2, 'double', {
+    'quotes': [2, 'single', {
       'avoidEscape': true,
       'allowTemplateLiterals': true
     }],
